requirements.txt
aio-pika==9.3.0
aiormq==6.7.7
alembic==1.12.1
async-timeout==4.0.3
asyncpg==0.29.0
greenlet==3.0.1
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
pamqp==3.2.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
yarl==1.9.2
